--- tests/ut/_310p/ops/test_fused_recurrent_gated_delta_rule_310.py
@@ -1,38 +1,0 @@
-import torch
-
-from vllm_ascend._310p.ops.fla.fused_recurrent_gated_delta_rule import fused_recurrent_gated_delta_rule_pytorch
-
-
-def test_fused_recurrent_gated_delta_rule_310_state_layout_matches_vllm():
-    q = torch.tensor([[[[1.0, 0.0]]]], dtype=torch.float32)
-    k = torch.tensor([[[[1.0, 0.0]]]], dtype=torch.float32)
-    v = torch.tensor([[[[10.0, 20.0, 30.0]]]], dtype=torch.float32)
-    g = torch.zeros(1, 1, 1, dtype=torch.float32)
-    beta = torch.ones(1, 1, 1, dtype=torch.float32)
-    initial_state = torch.tensor(
-        [[[[1.0, 2.0], [4.0, 8.0], [16.0, 32.0]]]],
-        dtype=torch.float32,
-    )
-
-    out, final_state = fused_recurrent_gated_delta_rule_pytorch(
-        q=q,
-        k=k,
-        v=v,
-        g=g,
-        beta=beta,
-        initial_state=initial_state,
-        inplace_final_state=False,
-        cu_seqlens=None,
-        ssm_state_indices=None,
-        num_accepted_tokens=None,
-        use_qk_l2norm_in_kernel=False,
-    )
-
-    expected_out = torch.tensor([[[[10.0, 20.0, 30.0]]]], dtype=torch.float32) / (2.0**0.5)
-    expected_state = torch.tensor(
-        [[[[10.0, 2.0], [20.0, 8.0], [30.0, 32.0]]]],
-        dtype=torch.float32,
-    )
-
-    torch.testing.assert_close(out, expected_out, rtol=1e-5, atol=1e-5)
-    torch.testing.assert_close(final_state, expected_state, rtol=1e-5, atol=1e-5)
